Support Gafaelfawr 2.0
Gafaelfawr 2.0, used for authentication in the Rubin Science Platform,
no longer uses JWTs for tokens and therefore the old token parsing
code no longer works.  Replace this with code that extracts user
information from the headers that are set by the NGINX auth request
handler.

Remove the code to get group membership and numeric UID information,
since neither should be necessary and don't appear to be used anywhere
else in the code base.

Add the additional Science Platform domains to the list of domains that
require authentication.