Remove .lsst.codes from sso.req.auth.hosts
This caused SUIT to include Gafaelfawr auth tokens in requests for
async results from the async-results.lsst.codes Google bucket,
which in turn confused Google and caused it to reject the request.
(The bucket does not require authentication.)

The intent of including .lsst.codes was to get SUIT to send auth
when making in-cluster requests for clusters in the .lsst.codes
domain, but this is primarily for TAP and we don't use SUIT with
TAP in clusters ending in .lsst.codes except for the minikube and
red-five test clusters.  Remove .lsst.codes again to fix the async
result retrieval, and we'll live with it not authenticating to TAP
in those clusters.  The important inclusion is .lsst.cloud, which
forces SUIT to send credentials for in-cluster requests in the IDF.