Stop attempting to get user's full name
We were passing the user's name in an HTTP header, but HTTP headers
only truly support ASCII (and, if not ASCII, use ISO 8859-1).  This
means we cannot represent many people's names.

Rather than use a more complicated encoding solution to pass the
name down, change SUIT to use the same behavior as squareone and only
show the user's username.  Do this by not setting the first or last
name attributes in UserInfo, which should trigger the Firefly
JavaScript layer to fall back to displaying only the username.